Update Django to always use Amazon SES email backend
To get this working, I configured the following in the AWS console:
1. Go to SES and select region US-East-1
2. In the sidebar, select Identity Management (Domains)
3. Click blue 'Verify new domain' button
4. Enter 'futurefeelslike.com' into the popup, check the DKIM box, and continue
5. On the next popup, click the 'Add to Route53' button at the bottom
6. Wait for the domain to verify
6. Back on the list of verified domains, select 'futurefeelslike.com' and then 'View Details'
7. On the details page, enable DKIM signing once the DKIM records are verified

Initially, new SES verified domains are in 'sandbox' mode. This means that all email addresses
you're sending to must be verified as well. I added each of the Urban Apps developer emails to
start, we can submit the request to Amazon support to remove the sandbox restriction at a later
date.

See:
http://docs.aws.amazon.com/ses/latest/DeveloperGuide/request-production-access.html