Update ClimateData view/serializer to serialize to new API spec
ClimateDataList view now returns filtered data (via PARAMS)
for a given city and scenario in the form:
```
{
  "city": { <City object> },
  "scenario": "RCP45",
  "models": ["model1", ... ],
  "variables": ["tasmax", ... ],
  "data": {
    "2001": {
       "taxmax": [273.15, ... ],
       ...
    },
    ...
  }
}
```
`models` is the list of models used to generate the AVG value for each day
  in the data object
`variables` is the list of variables provided in each of the `year` objects

Each year requested via the `years=...` query param is a key in the `data` object.

Removes old generate_climate_data factory method, replaces
with a mixin to generate a static set of climate data objects
for each tests.