Add logic to collate map data by cell for cities
This is intended to prevent duplicated data between cities that are
sufficiently close to fall within the same raw data at the source
resolution.

Cities reference their map cell by a foreign key relation... if a model
has different coordinates for its best match data point for a given
city, the daatabase schema will have to be changed to accomodate that.

Cities are matched with the closest data point coordinate using Manhattan
distance, comparing longitude and latitude independently. If this
becomes a problem we may need to adjust to using true distance
calculation.