Redirect from root to user profile
Adds a redirect from `/` to the user profile view, and adds
LoginRequiredMixin to that view so users who aren't logged in will be
further redirected to the login page.

Also, small changes to the URL config to remove redundancy and make it so
that arbitrary URLs under `account/` will be 404 rather than redirecting.