Remove front-heavy raw query for JIT check
Rather than loading a master list of all pairs of map cells and models
that already exist (Which required using a raw query in order to have
managable performance), this makes it so we check each map cell/model
pair as we import. The performance cost for each check is much lower and
should give a more consistent experience for users and should mean a
more stable design in case of model changes.