Add custom ObtainAuthToken view for Climate API
For now, the only customization on this view was to setup
a default throttling class, to limit requests to this endpoint.

This required enabling caching. The DRF docs state that for simple
cases Django's LocMemCache is sufficient, so that's what was chosen
until its demonstrated that we need something better.

There are many other things we could do to improve tokens, such
as adding expiration dates, or adding support for multiple tokens
per user. For now, this seemed like a sensible, simple balance.