Fix total precip indicator units confusion
Adding precip rates across days does the same thing as averaging them
then multiplying by the number of days, with the disadvantage that it
produces numbers that aren't in the units we think they are.

So this changes the yearly indicator to use an average and convert from
the storage unit to a yearly rate instead of converting to a daily rate
and adding up the day values.

It also adds the conversion functions for inches and sets the default
precip unit to inches per day.