Change remote import to skip null data points
Some models use a 365-day calendar, resulting in February 29th being
skipped each leap year and producing a null in the output. This, when
coupled with the fixed 366-day year output, meant that the remote import
could include many rows for days that did not exist in the source data.
This change introduces a filter to ensure that we do not create data
points for these non-modeled days.