Calculate extreme precipitation baseline on import
Extreme precipitation events are defined as surprassing 99th percentile
of historic rainfall. To project how these change over time, we need to
define what that baseline is for each area of concern.

This is done by processing historic data for each model independantly,
then averaging the results together. This is because variations in the
model's historic data causes a regression-to-the-mean effect, where some
models record some rainfall, others none, and on average every day
historically had some small amount of rainfall. By processing the models
independantly, our final result better matches actual real-world
recordings.

Importing baseline data is safe to restart and detects if data already
exists for the given city, so deleting it on startup isn't necessary. If
the loaded data needs refreshing the user can clear it manually.