Add Heat Wave Duration Index indicator
This involves a lot of changes under the hood. The climate data averages
model has been changed to represent a given map cell rather than city,
and tracks dates by day-of-year rather than month-date strings. The city
to map cell conversion is necessary because the indicator query works on
map cells and does not even know the city, and the date conversion
tremendously simplifies joining from the day-of-year climate data record.

Note that because of this change we need to clear all stored historic
average data.

This also includes refactoring some surrounding logic to be friendly
with generator indicators, and introduces a noop unit converter.