Convert CountIndicator query to Sum to catch zeroes
When we're trying to count days on which a condition is true, putting the
condition in the filter means we don't get rows when it's false, which
means we can't tell the difference between "you asked about this
year/model but it contains no matching days" and "you filtered out this
year/model in your query params".

For example, no matter what your filters, `data` will be empty in the
YearlyFrostDays response for Honolulu. Also the numbers will be skewed
for e.g. Austin, TX because the models that predict 0 frost days will
be ignored rather than averaged in.

This changes YearlyCountIndicator and MonthlyCountIndicator to instead
use a Sum(Case(When))) expression to count up days on which the
condition is true and properly return zero for periods in which it's
false.

This commit doesn't refactor the existing indicators that are doing the
counting in code in custom `aggregate` methods. Most of them are more
complicated (i.e. applying inter-day conditions like "consecutive dry
days").