Refactor time aggregation into base Indicator class
Time aggregation is now a parameter passed in during Indicator creation,
based on a list of allowed indicator aggregations. Some indicators only
support yearly or daily aggregation, but most support yearly or monthly.

This dramatically cuts down on the number of individual indicators, but
is a huge shift in the Indicator design. The final implementation is
likely to shift as we adjust for actual usage and try to incorporate
MaxConsecutiveDaysIndicator to being aggregation-agnostic.