Use memcached as cache backend for throttling
This commit only provides support for this in the dev environment,
support for AWS environment coming very soon.

Memcached chosen over LocMem cache so we have a global, fast cache,
rather than relying on a per-thread LocMem cache per gunicorn
worker, which will serve different throttled times and use more mem
because theres a separate cache per worker process.

Chose pylibmc because it has docs and an active maintainer, seems
simpler with more configuration available. python-memcached is
going through a maintainer switchup and it was hard to find docs
and info. Trivial to switch between the libraries if issues
occur down the line.