Add serializers, views for Region
Since ecoregions are distinct entities from cities, we give
them a top-level endpoint within the app.

The listing endpoint `/api/region/` lists the levels and
descriptions of all available regions. This has the same
searching, filtering, ordering, and paginating capabilities
as the `/api/city/` endpoint. The listing does not include
region geometries, as they can be quite heavy.

The detail endpoint `/api/region/{id}/` returns a GeoJSON
of the specified region. It includes levels and descriptions
as GeoJSON properties, as well as the actual geometry, usually
as a MultiPolygon.