Add SCSS imports example
Demonstrates how to use the @import directive with the
django-static-precompiler plugin.

Like regular SCSS, @imports are relative to the main file being
compiled. In addition, the plugin allows specifying extra folders
to pull files from, which can then be imported as if they were
also relative to the main file being compiled.