Restore development DB from an S3 backup
Our provisioning process sets up a database, but
doesn’t add any climate data to it. This commit modifies
`scripts/update` to restore the database from a backup
on S3. I also added the script to the Jenkinsfile to run
in CI, and updated the development setup and DB dump creation
instructions in the README. The logic to setup the database
without loading climate data has been moved to `scripts/setupdb`