Update to open registration in all cases
REGISTRATION_OPEN was set to false only if env.CC_REGISTRATION_OPEN='false',
and that's not set anywhere in code, Terraform or staging/prod tfvars, so
I have no idea how registration is being set to closed in those environments.

Either way, this should force registration to always be open. There's not much
of an advantage to setting this via ENV, because we'd need a deploy to change
registration status, which would be the same as flipping this flag in code.