Fix bugs introduced by refactor
- Cast lat/lon values to python floats from
  numpy.float32 in ClimateDataCell create
- Ensure that year and data_source.year are
  always both ints before attempting to compare
  in the assert check

Someone data_source.year is sometimes not an int,
even though the Django model specifies that it
should be. This occurs when do
ClimateDataSource.objects.create(year='2010', ...)
Until we refresh from db or request a new instance
with other queryset methods, data_source.year will
be a string value instead of the "cleaned" int.