Fix pending trigger events error in migration 0056
Due to the attempt to modify the table before the
change in RunPython commits via default
"CONSTRAINT DEFERRED" behavior

Solution from:
https://stackoverflow.com/questions/28429933/django-migrations-using-runpython-to-commit-changes

It would be better to split this migration, but this gets things working
again for now. There's no insert happening here, which is what could likely
cause trouble when we switch to immediately checking all constraints.