Support incremental nex2db inserts for new cities
This is done the easy way. We now always create or update
a ClimateDataYear record for a given map_cell + data_source

We no longer skip import for a given ClimateDataSource if
import_completed=True

For now this is an acceptible compromise, given that the changes
here update existing data in place without first deleting, and
we expect to make a few bulk additions of cities rather than
fully supporting user-triggered single city imports.