Add management command to update cities
Query geonames to get current data for an existing city in the database, and update it.
Fixes cities with unicode characters in their names being broken in the source data (s3 files).
Can also serve as a way to update population counts.
Also fixes five cities missing `admin` in the 16,000 city file (`admin` is the state abbreviation).

Fixes #765.