Move dynamic checking of which datasets a City supports to a DB field
The current dynamic check is not working very well. It checks for the
presence of map cells and reports their dataset names, but doesn't check
the values for those cells.
In practice, this means that a location without LOCA data will be reported
as supporting LOCA data, as the ingest will create LOCA cells, but fill
them with NaN (not a number) values.

Since checking the values in each cell would be computationally expensive,
we've instead moved the list of supported datasets to be stored in the
database. A future enhancement (#823) will ensure that this is properly
set for every existing city.

Closes #822