Include is_coastal in API views
Updates Lat/Lon API views to return a Feature
with properties instead of just a Geometry.

Re-adds City.is_coastal calculation in import_coastline:

We are leaving City.is_coastal in since its not trivial
to calculate via ClimateDataCells and we cannot use the
City centroid ClimateDataCell directly as a proxy for
whether a location is coastal or not since it excludes
places where the centroid is off the coast but the
city boundary intersects with the coastal geometry.